LayoutTests/sputnik/Conformance/12_Statement/12.6_Iteration_Statements/12.6.1_The_do_while_Statement/S12.6.1_A6_T1-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Expected token '('
S12.6.1_A6_T1

PASS Expected parsing failure

TEST COMPLETE
